+++ benchmark_test.go
@@ -6,56 +6,32 @@ import (
 	. "github.com/gbrlsnchs/patricia"
 )
 
-func BenchmarkSingleStatic(b *testing.B) {
+func BenchmarkLongString(b *testing.B) {
 	b.ReportAllocs()
 
-	tree := New("bench")
+	t := New("BenchmarkLongString")
 
-	tree.Add("/foo/bar/baz/qux", nil)
+	t.Add("This is a very, very long string, so let's benchmark it.", "foo")
 
 	for i := 0; i < b.N; i++ {
-		_ = tree.Get("/foo/bar/baz/qux")
+		_ = t.Get("This is a very, very long string, so let's benchmark it.")
 	}
 }
 
-func BenchmarkSingleDynamic(b *testing.B) {
+func BenchmarkManyWords(b *testing.B) {
 	b.ReportAllocs()
 
-	tree := New("bench")
+	t := New("BenchmarkManyWords")
 
-	tree.Add("/foo/:bar/:baz/:qux", nil)
-
-	for i := 0; i < b.N; i++ {
-		_, _ = tree.GetByRune("/foo/123/456/789", ':', '/')
-	}
-}
-
-func BenchmarkMultipleStatic(b *testing.B) {
-	b.ReportAllocs()
-
-	tree := New("bench")
-
-	tree.Add("/foo", nil)
-	tree.Add("/foo/bar", nil)
-	tree.Add("/foo/bar/baz", nil)
-	tree.Add("/foo/bar/baz/qux", nil)
+	t.Add("romane", 1)
+	t.Add("romanus", 2)
+	t.Add("romulus", 3)
+	t.Add("rubens", 4)
+	t.Add("ruber", 5)
+	t.Add("rubicon", 6)
+	t.Add("rubicundus", 7)
 
 	for i := 0; i < b.N; i++ {
-		_ = tree.Get("/foo/bar/baz/qux")
-	}
-}
-
-func BenchmarkMultipleDynamic(b *testing.B) {
-	b.ReportAllocs()
-
-	tree := New("bench")
-
-	tree.Add("/foo", nil)
-	tree.Add("/foo/:bar", nil)
-	tree.Add("/foo/:bar/:baz", nil)
-	tree.Add("/foo/:bar/:baz/:qux", nil)
-
-	for i := 0; i < b.N; i++ {
-		_, _ = tree.GetByRune("/foo/123/456/789", ':', '/')
+		_ = t.Get("romanus")
 	}
 }