+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2017-12-11
+### Added
+- Makefile.
+- Thread safety (if enabled).
+- Thread safety test and example.
+
+### Changed
+- Travis CI config script.
+
+### Fixed
+- `Node.IsLeaf` now returns a correct value.
+- This package now correctly implements a PATRICIA tree.
+
+### Removed
+- Travis CI script for running goimports.
+
 ## [0.3.0] - 2017-12-04
 ### Added
 - Alphabetical sorting.
@@ -72,6 +88,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[0.4.0]: https://github.com/gbrlsnchs/patricia/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/gbrlsnchs/patricia/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/gbrlsnchs/patricia/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/gbrlsnchs/patricia/compare/v0.1.3...v0.2.0